Refactor: 홈 화면 섹션 분리 #133
